fix #1, comment unused Slice, put Mixer on stack
This commit fixes the memory leaks as reported in #1, by removing
the unnecessary memory allocation for the Slice by #if 0 (it could
also be back-spaced, but I figured it might be useful for reference).

The Mixer is now allocated on the stack (just like a normal variable)
instead of on the heap using malloc(). Note that malloc() returns
un-initialized memory, so the contents could have been anything. Now we
create the Mixer on the stack (also un-initialized), but then call
memset(), which will zero the structure for us.

After that, the structure is now a normal variable, not a pointer, so
fix the references to that variable to be direct instead of pointer
accesses.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>